optimize waybill
